requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
Werkzeug==3.0.3
python-dotenv==1.0.1
